% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{function.enum}
\alias{function.enum}
\title{Function enum with mathematics functions}
\usage{
function.enum(type)
}
\arguments{
\item{type}{The function enum name}
}
\value{
Character list
}
\description{
Function enum with mathematics functions
}
