% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{insert}
\alias{insert}
\title{Insert value}
\usage{
insert(environmentId, document, key, value)
}
\arguments{
\item{environmentId}{The environnement who contains informations of one document}

\item{document}{The document who contains informations of multiple values}

\item{key}{The key of one value}

\item{value}{The value to insert}
}
\description{
Insert value
}
