% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{select}
\alias{select}
\title{Select value}
\usage{
select(environmentId, document, key)
}
\arguments{
\item{environmentId}{The environnement who contains informations of one document}

\item{document}{The document who contains informations of multiple values}

\item{key}{The key of one value}
}
\description{
Select value
}
