% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\docType{class}
\name{Storage}
\alias{Storage}
\title{Storage.}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}}.
}
\description{
Storage.

Storage.
}
\section{Methods}{

\describe{
  \item{\code{new(obj)}}{This method is used to create object of this class.}
  }
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Storage$new()}}
\item \href{#method-select}{\code{Storage$select()}}
\item \href{#method-insert}{\code{Storage$insert()}}
\item \href{#method-clone}{\code{Storage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Storage$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select"></a>}}
\subsection{Method \code{select()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Storage$select(environmentId, document, key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insert"></a>}}
\subsection{Method \code{insert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Storage$insert(environmentId, document, key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Storage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
