% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators.is.char.or.empty.list}
\alias{validators.is.char.or.empty.list}
\title{Used for validate if field is a vector of char OR a list}
\usage{
validators.is.char.or.empty.list(x)
}
\arguments{
\item{x}{The value to check}
}
\value{
Boolean
}
\description{
Used for validate if field is a vector of char OR a list
}
