% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators.is.empty}
\alias{validators.is.empty}
\title{Empty Value, use when receved a list}
\usage{
validators.is.empty(x)
}
\arguments{
\item{x}{an object to check its emptiness}
}
\description{
Empty Value, use when receved a list
}
\examples{
\dontrun{
validators.is.empty(NULL)     # [1] TRUE
validators.is.empty(list())     # [1] TRUE
validators.is.empty(NA)       # [1] TRUE
}
}
