% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators.is.guid}
\alias{validators.is.guid}
\title{Used for validate if field is guid format}
\usage{
validators.is.guid(string)
}
\arguments{
\item{string}{The character to check}
}
\value{
Boolean
}
\description{
Used for validate if field is guid format
}
