% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators.is.numeric.positive.or.zero}
\alias{validators.is.numeric.positive.or.zero}
\title{Used for validate if field is positive numeric >= 0}
\usage{
validators.is.numeric.positive.or.zero(num)
}
\arguments{
\item{num}{The numeric to check}
}
\value{
Boolean
}
\description{
Used for validate if field is positive numeric >= 0
}
