% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators.is.R6}
\alias{validators.is.R6}
\title{===================================
Can not use into function validator
===================================
Used for validate if object is class R6 object}
\usage{
validators.is.R6(obj, class)
}
\arguments{
\item{obj}{The R6 object to check}

\item{class}{The name of R6 class to check}
}
\description{
===================================
Can not use into function validator
===================================
Used for validate if object is class R6 object
}
