% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators.properties}
\alias{validators.properties}
\title{Used for validate the fields structures of JSON}
\usage{
validators.properties(
  obj,
  fieldname,
  required = FALSE,
  fieldtype = NULL,
  enum = NULL,
  nullpropertie = FALSE
)
}
\arguments{
\item{obj}{The object to check}

\item{fieldname}{The field name to try validate}

\item{required}{The value of field is required or not}

\item{fieldtype}{The value of field must be this type}

\item{enum}{The enum list who value of field must be find}

\item{nullpropertie}{This value of field must be null or not}
}
\description{
Used for validate the fields structures of JSON
}
